Renderer Init: Use correct types Use correct namespace

+++ source/2_platform/rex_opengl/include/rex_opengl/resources/buffer_layout.h
@@ -58,7 +58,7 @@ namespace rex
         BufferLayout(const BufferElements& elements);
 
         bool empty() const;
-        size_t size() const;
+        s32 size() const;
 
         const BufferElements& get_buffer_elements() const;
